server/requirements.txt
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 